Moved controller/rules/verify.py to utils/verify.py, fixed imports
