--- mapillary/controller/rules/__init__.py
@@ -1,1 +1,0 @@
-# Copyright (c) Facebook, Inc. and its affiliates. (http://www.facebook.com)
