+++ mapillary/utils/verify.py
